Moved Apple Watch Connection to Home Page
